stage-1: delivery — Auth/Users module with complete user system
- Implemented new entity structure (Users, UserIntroduce, UserRights, UserWallet)
- Created comprehensive authentication service with registration and login
- Updated DbContext with all related entities and proper relationships
- Built repository pattern for data access
- Added JWT authentication with proper token generation
- Created seed data generator with 500+ realistic test users
- Updated API endpoints and DTOs to match new schema
- All main projects building successfully (API + MVC)
- Ready for testing and integration

+++ src/GameCore.Domain/Interfaces/IUserWalletRepository.cs
@@ -7,5 +7,5 @@ public interface IUserWalletRepository
     Task<UserWallet?> GetByUserIdAsync(int userId);
     Task<UserWallet> CreateAsync(UserWallet wallet);
     Task<UserWallet> UpdateAsync(UserWallet wallet);
-    Task<decimal> GetBalanceAsync(int userId);
+    Task<int> GetBalanceAsync(int userId);
 }
